V4.3.1.1 - Ajustement

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->timestamps();
             $table->string('image_path');
             $table->boolean('is_banned')->default(0);            
-            $table->unsignedInteger('role_id')->default(1);
+            $table->unsignedTinyInteger('role_id')->default(1);
             $table->foreign('role_id')->references('id')->on('roles');
         });
         Schema::enableForeignKeyConstraints();
